Air drop mining count fixed

+++ src/Transactions/Controllers/Models/License.cs
@@ -6,7 +6,7 @@ namespace CodeRower.CCP.Controllers.Models
     {
         public Guid CustomerId { get; set; }
         public Guid LicenseId { get; set; }
-        public int Mined { get; set; }
+        public decimal Mined { get; set; }
         public string? Title { get; set; }
         public DateTime? ActivationDate { get; set; }
         public DateTime? ExpirationDate { get; set; }
